add audio node type
